changed firebase account, rectified appbar issue in view profile page
